Removed position dictionary
The position dictionary will be a great optimization to speed up interaction
with the map but at the moment it is too much to worry about.

This all arose because we encoded the position of a mesh as an ID.

This ID made it easy to create a dictionary that we could use to reference
all the meshes but due to the fact that there can be name collisions. I need
to change the IDs to be just a number instead of a position. This, unfortunately
had the effect of making the position dictionary useless.

I was thinking of another solution on how to do this, basically using the
X,Y,Z coordinates as keys in a sparse dictionary.

pos_dict[x][y][z] but that will have to wait.